Ajout de fonctionnalités + joueur ordinateur

+++ src/sample/Joueur.java
@@ -9,6 +9,6 @@ public class Joueur {
 
     public Liste Main = new Liste(); // Main du joueur
     public Liste Malus = new Liste(); // Malus du joueur
-    public int NbKilometre = 0;
+    public double NbKilometre = 0.0;
     public String id; // Nom du joueur
 }
